Rearranging files inside the branch
