+++ html/forget.html
@@ -1,0 +1,51 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <title>Forget Password</title>
+    <link rel="stylesheet" href="../css/LoginStyle.css" />
+    <link
+      rel="stylesheet"
+      href="http://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css"
+    />
+    <script type="module" src="../js/forget.js" defer></script>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  </head>
+  <body>
+    <div class="container">
+      <div class="back_arr" onclick="window.location.href='loginPage.html';">
+        <div></div>
+      </div>
+      <div class="cover">
+        <div class="front-panel">
+          <img src="../images/Chat-Icon-Transparent-Background.png" />
+          <div class="text">
+            <span class="text-1"> <br /> </span>
+            <span class="text-2"> </span>
+          </div>
+        </div>
+      </div>
+      <form class="px-20" action="#">
+        <div class="form-content">
+          <div class="forgot-password-form">
+            <div class="title-name">Forgot Password</div>
+            <div class="input-fields">
+              <div class="input-field">
+                <i class="fas fa-envelope"></i>
+                <input
+                  type="text"
+                  placeholder="Enter your email"
+                  id="email_forgot"
+                  required
+                />
+              </div>
+              <div class="button input-field">
+                <input type="submit" value="Submit" id="submit_forgot" />
+              </div>
+            </div>
+          </div>
+        </div>
+      </form>
+    </div>
+  </body>
+</html>
