--- tailwind-project/public/loginPage.html
@@ -1,107 +1,0 @@
-<!DOCTYPE html>
-
-<html lang="en" >
-  <head>
-    <meta charset="UTF-8">
-    <title> Login and Sign-up</title>
-    <link rel="stylesheet" href="LoginStyle.css">
-
-    <link rel="stylesheet" href="http://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css">
-    <script type="module" src="LoginPage.js" defer> </script>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-  </head>
-<body>
-<div class="container">
-  <input type="checkbox" id="flip">
-    <div class="cover">
-      
-      <div class="front-panel">
-        <img src="Chat-Icon-Transparent-Background.png" >
-        <div class="text">
-            <span class="text-1"> <br>  </span>
-            <span class="text-2"> </span>
-
-        </div>
-    </div>
-    <div class="back-panel">
-      <img src="Chat-Icon-Transparent-Background.png" >
-      <div class="text">
-        <span class="text-1"><br> </span>
-        <span class="text-2"></span>
-
-      </div>
- 
-</div>
-</div>
-  <form action="#">
-   <div class="form-content"> 
-    <div class="login-form">
-     <div class="title-name">Login</div>
-        <div class="input-fields">
-           <div class="input-field">
-               <i class="fas fa-envelope"></i>
-                 <input type="text" placeholder="Enter your email"  id="email_login" required>
-           </div>
-           <div class="input-field">
-            <i class="fas fa-lock"></i>
-              <input type="password" placeholder="Enter your password" id="password_login" required>
-        </div>
-        <div class="text"><a href="forget.html" id="forgot_password">Forgot password?</a></div>
-        <div class="button input-field">
-           
-              <input type="submit" value="Submit" id="submit_login">
-        </div>
-         <div class="text sign-up-text">New here? <label for="flip">Create an account here</label></div>
-        </div>
-
-    </div>
-    
-  </form>
-  
-  
-  <form action="#">
-    <div class="signup-form">
-        <div class="title-name">Sign-up</div>
-           <div class="input-fields">
-              <div class="role-container">
-                  
-                    
-                  <select class="role-select" id="role" required>
-                          <option value="" >--Choose Role--</option>
-                          <option value="user">User</option>
-                          <option value="admin">Admin</option>
-
-
-                    </select>
-                  
-              </div>
-              <div class="input-field">
-                <i class="fas fa-envelope"></i>
-                  <input type="text" placeholder="Enter your email" id="email" required>
-            </div>
-              <div class="input-field">
-               <i class="fas fa-lock"></i>
-                 <input type="password" placeholder="Enter your password" id="password" required>
-           </div>
-           <div class="button input-field">
-               
-            <input type="submit" value="Submit" id="submit_register">
-           </div>
-           <div class="text sign-up-text">Already a member? <label for="flip">Login here</label></div>
-           </div>
-   
-       </div>
-   
-
-    </div>
-  </form>
-
-</div>
-
-
-
-</body>
-
-
-</html>
